refactor(omegaform): extract errors + HOC helpers
