+++ packages/vue-components/src/components/OmegaForm/useOmegaForm.ts
@@ -4,8 +4,9 @@
 import { DeepValue, type FormAsyncValidateOrFn, type FormValidateOrFn, type StandardSchemaV1, useForm } from "@tanstack/vue-form"
 import { Array, Context, Effect, Order, S } from "effect-app"
 import { UnionToTuples } from "effect-app/utils"
-import { Component, computed, ComputedRef, ConcreteComponent, h, type InjectionKey, onUnmounted, Ref, ref, watch } from "vue"
-import { useIntl } from "../../utils"
+import { ComputedRef, type InjectionKey, ref, watch } from "vue"
+import { eHoc, makeFieldMap } from "./errors"
+import { fHoc } from "./hocs"
 import { MergedInputProps } from "./InputProps"
 import OmegaArray from "./OmegaArray.vue"
 import OmegaAutoGen from "./OmegaAutoGen.vue"
@@ -20,6 +21,7 @@ import { FormErrors, makeSubmitHandlers, wrapOnSubmit } from "./submit"
 import { makeRunPromise } from "@effect-app/vue/runtime"
 
 export { FormErrors } from "./submit"
+export { useErrorLabel } from "./errors"
 
 type keysRule<T> =
   | {
@@ -30,127 +32,6 @@ type keysRule<T> =
     keys?: "You should only use one of banKeys or keys, not both, moron"
     banKeys?: NestedKeyOf<T>[]
   }
-
-const fHoc = (form: OF<any, any>) => {
-  return function FormHoc<P>(
-    WrappedComponent: Component<P>
-  ): ConcreteComponent<P> {
-    return {
-      render() {
-        return h(WrappedComponent, {
-          form,
-          ...this.$attrs
-        } as any, this.$slots)
-      }
-    }
-  }
-}
-
-export const useErrorLabel = (form: OF<any, any>) => {
-  const { formatMessage } = useIntl()
-  const humanize = (str: string) => {
-    return str
-      .replace(/([A-Z])/g, " $1") // Add space before capital letters
-      .replace(/^./, (char) => char.toUpperCase()) // Capitalize the first letter
-      .trim() // Remove leading/trailing spaces
-  }
-  const fallback = (propsName: string) =>
-    formatMessage
-      ? formatMessage({ id: `general.fields.${propsName}`, defaultMessage: humanize(propsName) })
-      : humanize(propsName)
-  const i18n = (propsName: string) =>
-    form.i18nNamespace
-      ? formatMessage({ id: `${form.i18nNamespace}.fields.${propsName}`, defaultMessage: fallback(propsName) })
-      : fallback(propsName)
-
-  return i18n
-}
-
-const eHoc = (errorProps: {
-  form: OF<any, any>
-  fieldMap: Ref<Map<string, { id: string; label: string }>>
-}) => {
-  return function FormHoc<P>(
-    WrappedComponent: Component<P>
-  ): ConcreteComponent<P> {
-    return {
-      setup() {
-        const { fieldMap, form } = errorProps
-        const generalErrors = form.useStore((state) => state.errors)
-        const fieldMeta = form.useStore((state) => state.fieldMeta)
-        const errorMap = form.useStore((state) => state.errorMap)
-
-        const errorLabel = useErrorLabel(form)
-
-        const errors = computed(() => {
-          // Collect errors from fieldMeta (field-level errors for registered fields)
-          const fieldErrors = Object.entries(fieldMeta.value).reduce<OmegaError[]>((acc, [key, m]) => {
-            const fieldErrors = (m as { errors?: Array<{ message?: string }> } | undefined)?.errors ?? []
-            if (!fieldErrors.length) {
-              return acc
-            }
-
-            const fieldInfo = fieldMap.value.get(key)
-            if (!fieldInfo) {
-              return acc
-            }
-
-            acc.push({
-              label: fieldInfo.label,
-              inputId: fieldInfo.id,
-              errors: [fieldErrors[0]?.message].filter(Boolean) as string[]
-            })
-
-            return acc
-          }, [])
-
-          // Collect errors from errorMap.onSubmit ONLY for fields that are NOT registered
-          // (registered fields already have their errors in fieldMeta)
-          const submitErrors: OmegaError[] = []
-          if (errorMap.value.onSubmit) {
-            for (const [_, issues] of Object.entries(errorMap.value.onSubmit)) {
-              if (Array.isArray(issues) && issues.length) {
-                for (const issue of issues) {
-                  const issAny: any = issue
-                  if (issAny?.path && Array.isArray(issAny.path) && issAny.path.length) {
-                    // Use the path from the issue to identify the field
-                    const fieldPath = issAny.path.join(".")
-                    // Only add errors for fields that are NOT registered (not in fieldMap)
-                    // Registered fields will already have their errors from fieldMeta
-                    if (!fieldMap.value.has(fieldPath)) {
-                      submitErrors.push({
-                        label: errorLabel(fieldPath),
-                        inputId: fieldPath,
-                        errors: [issAny.message].filter(Boolean)
-                      })
-                      // Only show first error per field, so break after adding
-                      break
-                    }
-                  }
-                }
-              }
-            }
-          }
-
-          // Combine both error sources (no need to check for duplicates since they're mutually exclusive)
-          return [...fieldErrors, ...submitErrors]
-        })
-
-        return {
-          generalErrors,
-          errors
-        }
-      },
-      render({ errors, generalErrors }: any) {
-        return h(WrappedComponent, {
-          errors,
-          generalErrors,
-          ...this.$attrs
-        } as any, this.$slots)
-      }
-    }
-  }
-}
 
 export type Policies = PersistencyPolicies
 export type defaultValuesPriorityUnion = PersistencyPriority
@@ -733,7 +614,7 @@ export const useOmegaForm = <
 
   const { handleSubmit, handleSubmitEffect } = makeSubmitHandlers<From, To>(form)
 
-  const fieldMap = ref(new Map<string, { label: string; id: string }>())
+  const { fieldMap, registerField } = makeFieldMap()
 
   const formWithExtras: OF<From, To> = Object.assign(form, {
     i18nNamespace: omegaConfig?.i18nNamespace,
@@ -744,20 +625,7 @@ export const useOmegaForm = <
     handleSubmit,
     // /** @experimental */
     handleSubmitEffect,
-    registerField: (field: ComputedRef<{ name: string; label: string; id: string }>) => {
-      watch(field, (f) => {
-        fieldMap.value.set(f.name, { label: f.label, id: f.id })
-      }, { immediate: true })
-      onUnmounted(() => {
-        // Only delete if we still own this entry (id matches)
-        // This prevents old components from deleting entries registered by new components
-        // during re-mount transitions (e.g., when :key changes)
-        const currentEntry = fieldMap.value.get(field.value.name)
-        if (currentEntry?.id === field.value.id) {
-          fieldMap.value.delete(field.value.name)
-        }
-      })
-    }
+    registerField
   })
 
   // Clear all field onSubmit errors when any value changes after a failed submission.
